refactor: comprehensive codebase cleanup and decoupling
- Decouple large `use-conversation-store.ts` into smaller modules in `src/stores/conversation/`.
- Centralize `charsPerToken` constant and apply DRY in token estimation logic.
- Refactor base components and icons for consistency.
- Improve naming and remove redundant comments across the `src/` directory.
- Add `VITE_APP_URL` to environment validation.
- Fix functional regression in conversation creation logic.
- Enhance type safety by minimizing `as any` and favoring explicit types.
- Ensure all components and services follow project-wide architectural patterns.
