+++ src/constants.ts
@@ -1,9 +1,8 @@
-export const lottieBaseSpinnerUrl =
-  'https://lottie.host/61eb14b2-5dd2-471a-88c3-9e9c61862e83/Rldo3dbFyi.lottie'
-
 export const defaultModel = 'gemini:gemini-2.5-flash'
 
 export const streamingCommitIntervalMs = 2_000
+
+export const charsPerToken = 4
 
 export const socketTimeoutMs = {
   emit: Number(15_000),
