+++ src/stores/conversation/use-conversation-list-state.ts
@@ -1,14 +1,14 @@
 import { ref } from 'vue'
-import { ok, err, type Result } from 'neverthrow'
+import { ok, err, type Result, ResultAsync } from 'neverthrow'
 import type { Conversation } from '@/database/types'
 import type { AppDb } from '@/database/app-db'
-import type { DbError } from '@/errors'
-import { ChatError } from '@/errors'
+import { DbError, ChatError } from '@/errors'
 
 export function useConversationListState(appDb: AppDb) {
   const conversations = ref<Conversation[]>([])
   const isLoadingConversations = ref(false)
   const messageCountsByConversation = ref<Map<number, number>>(new Map())
+  const error = ref<string | undefined>(undefined)
 
   let creationPromise: Promise<Result<number, DbError | ChatError>> | undefined
 
@@ -52,16 +52,19 @@ export function useConversationListState(appDb: AppDb) {
           isActive: 1,
           createdAt: new Date(),
           updatedAt: new Date(),
-        } as Conversation)
+        })
 
-        if (addResult.isErr()) return err(addResult.error)
+        if (addResult.isErr()) {
+          error.value = 'Failed to create conversation'
+          return err(addResult.error)
+        }
 
         const newId = Number(addResult.value)
         const loadResult = await loadConversations()
 
         if (loadResult.isErr()) {
           await appDb.conversation.delete(newId)
-          return err(new ChatError('Failed to load after creation', loadResult.error))
+          return err(new ChatError('Failed to load conversations after creation', loadResult.error))
         }
 
         if (!conversations.value.some((c) => c.id === newId)) {
@@ -78,13 +81,65 @@ export function useConversationListState(appDb: AppDb) {
     return creationPromise
   }
 
+  async function updateConversationTitle(
+    conversationId: number,
+    title: string
+  ): Promise<Result<void, DbError>> {
+    const result = await ResultAsync.fromPromise(
+      appDb.conversation.update(conversationId, {
+        title,
+        updatedAt: new Date(),
+      }),
+      (unknownError) => {
+        if (unknownError instanceof Error) {
+          return new DbError(unknownError.message, unknownError)
+        }
+        return new DbError('Failed to update conversation title', unknownError)
+      }
+    )
+
+    if (result.isErr()) {
+      error.value = 'Failed to update conversation title'
+      return err(result.error)
+    }
+
+    const conversation = conversations.value.find((conv) => conv.id === conversationId)
+    if (conversation) {
+      conversation.title = title
+      conversation.updatedAt = new Date()
+    }
+
+    error.value = undefined
+    return ok(undefined)
+  }
+
+  async function deleteConversation(
+    conversationId: number
+  ): Promise<Result<void, DbError | ChatError>> {
+    const result = await appDb.chatRepository.deleteConversationWithMessage(conversationId)
+
+    if (result.isErr()) {
+      error.value = 'Failed to delete conversation'
+      return err(result.error)
+    }
+
+    const loadResult = await loadConversations()
+    if (loadResult.isErr()) return err(loadResult.error)
+
+    error.value = undefined
+    return ok(undefined)
+  }
+
   return {
     conversations,
     isLoadingConversations,
     messageCountsByConversation,
+    error,
     loadConversations,
     loadMessageCounts,
     getMessageCount,
     createConversation,
+    updateConversationTitle,
+    deleteConversation,
   }
 }
